% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countrynames.R
\name{subsaharaCountries}
\alias{subsaharaCountries}
\title{Dirichlet-sample of a multinomial election poll}
\usage{
subsaharaCountries()
}
\value{
A character vector with subsaharan country names.
}
\description{
Description
}
\section{Warning}{

Do not operate heavy machinery within 8 hours of using this function.
}

\examples{
subsahara()
}
