% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countrynames.R
\name{unifyCountrynames}
\alias{unifyCountrynames}
\title{Unify countrynames}
\usage{
unifyCountrynames(inVec)
}
\arguments{
\item{inVec}{A vector or character string of country name(s)}
}
\value{
A character vector containing the unified strings.
}
\description{
Description
}
\section{Warning}{

Do not operate heavy machinery within 8 hours of using this function.
}

\examples{
x = c("Congo Kinshasa",
      "Congo, Republic of",
      "Congo Brazzaville",
      "Cote d`Ivoire",
      "Central African Rep.",
      "Cabo Verde",
      "Gambia, The")
      
 unifyCountrynames(x)
}
